chore: moved to module `builder` and implement `AddAssign` for `Builder`
